requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.0.0
psycopg2-binary>=2.9.0
Pillow>=9.0.0
reportlab>=3.6.0,<4.0
gunicorn>=20.1.0
whitenoise>=6.0.0
dj-database-url>=2.0.0
